Starting zoning of other flogiston models.
Right now this expects to work with the CMS and Asset plugins, and the
installation template has been updated to require them.

Starting the Flogiston::Zoning extension module, which will contain all
the behavior needed for the various models. Tricked out some wacky
auto-extension in the installation template. And, of course, started
specs for the various models detailing their new behavior. The
association is the easy stuff. On to more fun.